requirements.txt
ibm-cos-sdk==2.13.6
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
python-dotenv==1.0.0
